Add feature for send the m3u8 and the segments to object storage

+++ domain/xyzvod/vod_query.go
@@ -11,7 +11,7 @@ type Vod struct{
     Title string `json:"title"`
     Description string `json:"description"`
     Duration int `json:"duration"`
-    Created int `json:"created"`
+    Created int64 `json:"created"`
 }
 
 type VodQuery struct{}
